==> notebooks/01_load_and_validate_raw_data.ipynb <==
### Notebook for data load and EDA

# libs import
from pathlib import Path
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os
import sys
import warnings
import lux
import tqdm
import requests
from typing import List, Optional
import plotly.express as px

# remove warnings
warnings.filterwarnings('ignore')

# retina display -> delete this line if you don't have a retina display
%config InlineBackend.figure_format = 'retina'

# data path and import 
import requests

def download_one_file_of_raw_data(year: int, month: int) -> Path:
    """Downloads one file of raw data from the website of the TLC Trip Record Data and saves it to the data/raw folder.
    Args:
        year (int): The year of the data to download.
        month (int): The month of the data to download.
    Returns:
        Path: The path to the downloaded file.
    """
    # create the url
    url = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet"
    response = requests.get(url)

    if response.status_code == 200:
        # create the path to the file
        path = f'../data/raw/rides_{year}-{month:02d}.parquet'
        open(path, 'wb').write(response.content)
        return path #return Path(path) 
    else:
        print(f"Could not download file {url}")
        return None


# call the function
download_one_file_of_raw_data(2022, 1)

# read the file into a pandas dataframe
rides = pd.read_parquet('../data/raw/rides_2022-01.parquet')
rides.head()

# we need to know the pickup time and drop off location
rides = rides[['tpep_pickup_datetime', 'PULocationID']]

# rename the columns
rides.rename(columns={
    'tpep_pickup_datetime': 'pickup_datetime', 
    'PULocationID': 'pickup_location_id'}, 
    inplace=True)
#rides.head()

# describe the data
rides.pickup_datetime.describe()

# remove the rides that are not in 2022
# we need only Jan 2022
rides = rides[rides.pickup_datetime >= '2022-01-01']
rides = rides[rides.pickup_datetime <  '2022-02-01']
rides.pickup_datetime.describe()

# save the data
rides.to_parquet('../data/transformed/validated_rides_2022-01.parquet')

# read transformed data
rides = pd.read_parquet('../data/transformed/validated_rides_2022-01.parquet')
rides.head()
