==> notebooks/08_lightgbm_model.ipynb <==
## Using LGBM
To Do's
- add visualizations
- add feature importance  [done] 
    - for 4 weeks of data [done] 
    - for 8 weeks of data [done] 
- test 8 weeks of data [done] 

### 4 weeks of data

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2022, 6, 1, 0, 0, 0),
    target_column_name='target_rides_next_hour'
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

import lightgbm as lgb

# use only past rides data - input features
past_rides_columns = [col for col in X_train.columns if col.startswith('rides_')]
X_train_only_numeric = X_train[past_rides_columns]

# train model
model = lgb.LGBMRegressor()
model.fit(X_train_only_numeric, y_train)

X_test_only_numeric = X_test[past_rides_columns]
predictions = model.predict(X_test_only_numeric)
predictions

from sklearn.metrics import mean_absolute_error
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=}')

### Feature Importance for 4 weeks of data

# feature importance
import numpy as np

feature_importance = pd.DataFrame(
    model.feature_importances_,
    index=X_train_only_numeric.columns,
    columns=['importance']
)
feature_importance.sort_values('importance', ascending=False, inplace=True)
feature_importance.head(20) 

TOP 3
- rides_previous_1_hour	  = 224
- rides_previous_335_hour = 58
- rides_previous_504_hour = 42

# use only important features in the model and retrain
important_features = feature_importance[feature_importance['importance'] > 10].index # for XGBOOST > 0.001
X_train_important = X_train[important_features]
X_test_important = X_test[important_features]

model_important = lgb.LGBMRegressor()
model_important.fit(X_train_important, y_train)

important_features

important_features.size, X_train_important.shape[1], X_test_important.shape[1]

# evaluate the model
predictions_important = model_important.predict(X_test_important)
test_mae_important = mean_absolute_error(y_test, predictions_important)
print(f'{test_mae_important=}')

### Summary 4 weeks 
[w\o feature importance] 
- mae = 2.57
- feature_size = 675
- train time = 5 sec

[with feature importance] 
- mae_important = 2.54
- feature_size = 33
- train time = 2 sec

### 8 weeks of data

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data_2_months.parquet')
df

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2022, 6, 1, 0, 0, 0),
    target_column_name='target_rides_next_hour'
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

# use only past rides data
past_rides_columns = [col for col in X_train.columns if col.startswith('rides_')]
X_train_only_numeric = X_train[past_rides_columns]

# train model
model = lgb.LGBMRegressor()
model.fit(X_train_only_numeric, y_train)

X_test_only_numeric = X_test[past_rides_columns]
predictions = model.predict(X_test_only_numeric)
predictions

from sklearn.metrics import mean_absolute_error
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=}')

### Feature Importance for 8 weeks of data

# feature importance
import numpy as np

feature_importance = pd.DataFrame(
    model.feature_importances_,
    index=X_train_only_numeric.columns,
    columns=['importance']
)
feature_importance.sort_values('importance', ascending=False, inplace=True)
feature_importance.head(20) 

# use only important features in the model and retrain
important_features = feature_importance[feature_importance['importance'] > 10].index # for XGBOOST > 0.001
X_train_important = X_train[important_features]
X_test_important = X_test[important_features]

model_important = lgb.LGBMRegressor()
model_important.fit(X_train_important, y_train)

important_features.size, X_train_important.shape[1], X_test_important.shape[1]

# evaluate the model
predictions_important = model_important.predict(X_test_important)
test_mae_important = mean_absolute_error(y_test, predictions_important)
print(f'{test_mae_important=}')

### Summary 8 weeks 
[w\o feature importance] 
- mae = 2.74
- feature_size = 1347
- train time = 7.5 sec

[with feature importance] 
- mae_important = 2.66 
- feature_size = 24
- train time = 1.5 sec


## Experiments section
To Do:
- [] Experiment with 16 weeks (4 months) of data - to get the 70\30 split of the train and test data
