==> notebooks/11_backfill_feature_store.ipynb <==
HOPSWORKS_PROJECT_NAME = "taxi_demand_bormaley"

import os
from dotenv import load_dotenv
from src.paths import PARENT_DIR

load_dotenv(PARENT_DIR/ '.env')
HOPSWORKS_API_KEY=os.environ['HOPSWORKS_API_KEY']

from datetime import datetime
import pandas as pd
from src.data import load_raw_data

# load data for the specified years
from_year = 2022
to_year = datetime.now().year
print(f'Downloading raw data from {from_year} to {to_year}')

rides = pd.DataFrame()
for year in range(from_year, to_year + 1):

    # download data for the entire year (all months)
    rides_one_year = load_raw_data(year)

    # append rows
    rides = pd.concat([rides, rides_one_year])

print(f'{len(rides)} rows downloaded')

from src.data import transform_raw_data_into_ts_data

# transform the data into ts data
ts_data = transform_raw_data_into_ts_data(rides)

# save data into hopsworks feature store
# pip install setuptools (upgrade if needed)) 
# pip install great_expectations (upgrade if needed)

import hopsworks as hs

# connect to the feature store
project = hs.login(
    project=HOPSWORKS_PROJECT_NAME,
    api_key_value=HOPSWORKS_API_KEY,
)

# get the feature store 
feature_store = project.get_feature_store()

# define the feature group
FEATURE_GROUP_NAME = 'time_series_hourly_feature_group'
FEATURE_GROUP_VERSION = 1

# create the feature group
feature_group = feature_store.get_or_create_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION,
    description='Time series data at hourly frequency',
    primary_key=['pickup_location_id', 'pickup_hour'],
    event_time='pickup_hour',
)

# save the data into the feature store
feature_group.insert(ts_data, write_options={"wait_for_job": False})

# test
import hopsworks as hs
